requirements.txt
ogb
pyg-lib
torch-scatter
torch-sparse
torch-cluster
torch-spline-conv
torch-geometric
